Fix bugs
Move window to the last recieved ACK, not resend all the packets if one of the ACKs was missing.
Change mod to window_size + 1, to handle the case when no packets in the windows recieved thier ACKs.
Resend the ACK for next expected packet in reciever even if ignoring the recieved packet.