Send only new NACKs or by timer
The transmission was choking: there were too many NACKs sent. Sender sent the whole
window but keeps getting NACKs for previous one. As time goes on, the number of NAKCs
for previous windows keeps growing.

To prevent that, send only new NACKs: if this NACK was sent, don't send it again, until
the expected_seq_num is changed. Or if the current NACK timeout was reached, in case NACK
was lost in transmission